requirements.txt
fastapi
pydantic-settings
httpx
uvicorn[standard]
fastapi-cache2
Jinja2